product feature: refactor get product to become reusable
